Add weapon proficiency system and assorted fixes
Introduce weapon proficiency support and related helpers, plus multiple bugfixes and refactors.

- Add proficiency helper script (createProficiencyCatalyst) and replace duplicated catalyst actions.
- Add weapon proficiency config keys and enums; load settings via ConfigManager.
- Implement/extend WeaponProficiency: JSON loading, KV (de)serialization, experience arrays as vectors, perk/level limits, safer parsing, various logic fixes (boss/forge checks, bestiary clamping, mastery calculations), skill percentage mapping, and API tweaks (saveAll returns bool).
- Add new proficiency JSON loader error handling to throw on failure.
- Fix save_interval global event: robust time parsing, scheduling, broadcasting and return values.
- Imbuements: map scroll IDs for fast lookup, fix parsing, reload handling, and ordering when applying scrolls; small gameplay fix for imbuement failure flow.
- Items XML: correct augments count and add primarytype/article attributes for boots.
- Fix typos (aditionalTarget -> additionalTarget) across wheel and spell augment code and related header changes.
- UI/UX and safety fixes: feedback message for weapon_proficiency god talk action and etcher cancel message.
- Misc: small includes and header adjustments, logging improvements, and various defensive checks to avoid crashes.

These changes improve weapon proficiency functionality, remove duplicated code, and harden several systems for stability and correctness.

+++ src/items/weapons/weapons.cpp
@@ -132,7 +132,7 @@ void Weapon::configureWeapon(const ItemType &it) {
 	id = it.id;
 }
 
-int32_t Weapon::playerWeaponCheck(const std::shared_ptr<Player> &player, const std::shared_ptr<Creature> &target, uint8_t shootRange) const {
+int32_t Weapon::playerWeaponCheck(const std::shared_ptr<Player> &player, const std::shared_ptr<Creature> &target, int32_t shootRange) const {
 	shootRange += player->weaponProficiency().getStat(WeaponProficiencyBonus_t::ATTACK_RANGE);
 
 	const Position &playerPos = player->getPosition();
